✅ Thumbnail auto removal if set
The issues submitted by someone that thumbnail was applying to only one video (if set) fixed that ✅✅
Now it will preserve and will apply to all files( if set)

+++ devgagan/core/get_func.py
@@ -181,7 +181,8 @@ async def upload_media(sender, target_chat_id, file, caption, edit, topic_id):
 
     finally:
         if thumb_path and os.path.exists(thumb_path):
-            os.remove(thumb_path)
+            if os.path.basename(thumb_path) != f"{sender}.jpg":  # Check if the filename is not {sender}.jpg
+                os.remove(thumb_path)
         gc.collect()
 
 
@@ -671,7 +672,7 @@ async def callback_query_handler(event):
         await event.respond('Please send the photo you want to set as the thumbnail.')
     
     elif event.data == b'pdfwt':
-        await event.respond("Watermark is Pro+ Plan.. contact @kingofpatal")
+        await event.respond("This feature is not available yet in public repo...")
         return
 
     elif event.data == b'uploadmethod':
